Bring up to date with PCSX-R master (97809)

+++ libpcsxcore/cdrom.c
@@ -1328,7 +1328,7 @@ void cdrWrite1(unsigned char rt) {
 			set_loc[i] = btoi(cdr.Param[i]);
 
 		i = msf2sec(cdr.SetSectorPlay);
-		i = abs(i - msf2sec(set_loc));
+		i = abs(i - (int)msf2sec(set_loc));
 		if (i > 16)
 			cdr.Seeked = SEEK_PENDING;
